libcamera: software_isp: Use floating point for color parameters
It's more natural to represent color gains and black level as floating
point numbers rather than using a particular pixel-related
representation.

double is used rather than float because it's a more common floating
point type in libcamera algorithms.  Otherwise there is no obvious
reason to select one over the other here.

The constructed color tables still use integer representation for
efficiency.

+++ src/ipa/simple/ipa_context.h
@@ -20,7 +20,7 @@ namespace ipa::soft {
 struct IPASessionConfiguration {
 	float gamma;
 	struct {
-		uint8_t level;
+		double level;
 		bool set;
 		bool changed;
 	} black;
@@ -28,9 +28,9 @@ struct IPASessionConfiguration {
 
 struct IPAActiveState {
 	struct {
-		unsigned int red;
-		unsigned int green;
-		unsigned int blue;
+		double red;
+		double green;
+		double blue;
 	} gains;
 	static constexpr unsigned int kGammaLookupSize = 1024;
 	std::array<double, kGammaLookupSize> gammaTable;
